requirements.txt
streamlit
google-cloud-firestore
mailjet_rest
openai